Thursday (2/3/2022) -Further Creation of the ViewPager, Bug Fixes, Constraint Fixes

+++ app/src/main/java/com/maxwell/recyclerview/adapter/CardAdapter.kt
@@ -36,7 +36,7 @@ class CardAdapter(private val context: Context, private val myModelArrayList: Ar
         view.cardNumber.text = bcardnumber
         view.cardExpiry.text = bcardexpiry
         view.cardName.text = bcardname
-        view.cardColor.setBackgroundColor(bcardcolor.toInt())
+        view.cardColor.setBackgroundColor(bcardcolor)
 
         view.setOnClickListener {
             Toast.makeText(context, "$bcardname \n $bcardholder \n $bcardnumber", Toast.LENGTH_SHORT).show()
